@@ -1,48 +1,49 @@
-import ImprovementCard from "@/components/resumes/cards/improvement";
+"use client";
+
+import Feedback from "@/components/resumes/feedback";
+import { context } from "@/lib/context";
 import Image from "next/image";
+import { useState } from "react";
 
 export default function Page({ params }: { params: { slug: string } }) {
+  const { slug } = params;
+  const [status, setStatus] = useState<"Done" | "Loading" | "Analyzing" | "Analyzed" | "Saving to database">(
+    "Loading"
+  );
   return (
-    <div className="p-4 md:p-0 w-full flex flex-col">
-      <h1 className="mt-[-6px] text-4xl text-black">
-        Improve your resume
-      </h1>
-      <p className="text-gray-400">
-        Get specific, actionable feedback on how to improve your resume. The Al
-        will highlight areas for improvement and provide clear recommendations
-        to help you make your resume stand out for your next job application.
-      </p>
+    <>
+      <context.resume.ChangeTitle title="Improve your resume" />
+      <div className="p-4 md:p-0 w-full flex flex-col">
+        <p className="text-gray-400">
+          Get specific, actionable feedback on how to improve your resume. The
+          Al will highlight areas for improvement and provide clear
+          recommendations to help you make your resume stand out for your next
+          job application.
+        </p>
 
-      <div className="mt-8 flex flex-col md:flex-row justify-betwee">
-        <div className="w-full self-center">
-          <p className="hover:cursor-pointer hover:text-gray-500 underline text-gray-400">
-            EliasWambuguResume.pdf
-          </p>
-          <p className="font-bold">Your resume is ready for review!</p>
-          
-        </div>
-        <Image
-          className="bg-none"
-          height={100}
-          width={300}
-          src="/images/resume.jpg"
-          alt=""
-        />
-      </div>
-      <h2 className="mt-44 font-bold text-2xl">Suggested Improvements</h2>
-          <p className="mt-4">
-            {"We've used AI to help you improve your resume!"}
-          </p>
-      <div className="flex flex-col mt-8">
-        {[...Array(3)].map((_, i) => (
-          <ImprovementCard
-            key={i}
-            title="Add a summary section"
-            description="A summary section can also help you tailor your resume to specific jobs by highlighting the most relevant experience and skills.
-          Add a summary section at the top of your resume. This should be a concise paragraph that summarizes your career history, skills, and accomplishments. It's a great way to grab the reader's attention and give them a quick overview of who you are and what you're looking for."
+        <div className="mt-8 flex flex-col md:flex-row justify-betwee">
+          <div className="w-full self-center">
+            <p className="hover:cursor-pointer hover:text-gray-500 underline text-gray-400">
+              EliasWambuguResume.pdf
+            </p>
+            <p className="font-bold">
+              {status === "Done" ? "Your feedback is ready!" : status + "..."}
+            </p>
+          </div>
+          <Image
+            className="bg-none"
+            height={100}
+            width={300}
+            src="/images/resume.jpg"
+            alt=""
           />
-        ))}
+        </div>
+        <h2 className="mt-44 font-bold text-2xl">Suggested Improvements</h2>
+        <p className="mt-4 mb-8">
+          {"We've used AI to help you improve your resume!"}
+        </p>
+        <Feedback status={status} setStatus={setStatus} slug={slug} />
       </div>
-    </div>
+    </>
   );
 }
