@@ -1,4 +1,5 @@
-import { Button } from "@/components/ui/button";
+
+import UploadButton from "@/components/upload";
 
 export default function Page() {
   return (
@@ -13,20 +14,9 @@ export default function Page() {
           }
         </p>
         <p className="font-bold text-xl mt-8">Upload your resume</p>
-        <div className="mt-8  flex items-center justify-center w-full">
-          <label className="flex flex-col items-center justify-center w-full h-64 border-2 border-gray-300 border-dashed rounded-lg cursor-pointer dark:hover:bg-bray-800 dark:bg-gray-700 hover:bg-gray-50 dark:border-gray-600 dark:hover:border-gray-500 dark:hover:bg-gray-600">
-            <div className="flex flex-col items-center justify-center pt-5 pb-6">
-              <p className="mb-2 text-sm font-bold text-black dark:text-gray-400">
-                Drag and drop your resume here
-              </p>
-              <div className="p-4 text-sm rounded-lg flex justify-center items-center text-white h-10 bg-black">
-                or browse files
-              </div>
-            </div>
-            <input id="dropzone-file" type="file" className="hidden" />
-          </label>
-        </div>
-        <p className="mt-2">We support .doc, .docx, and .pdf files.</p>
+        
+        <UploadButton endpoint="resumeUpload" />
+        <p className="mt-2">We support .pdf files.</p>
       </div>
     </div>
   );
